add TDatabaseObjectToggle trait
see #2728
