+++ wcfsetup/install/files/lib/data/captcha/question/CaptchaQuestionAction.class.php
@@ -2,6 +2,7 @@
 namespace wcf\data\captcha\question;
 use wcf\data\AbstractDatabaseObjectAction;
 use wcf\data\IToggleAction;
+use wcf\data\TDatabaseObjectToggle;
 
 /**
  * Executes captcha question-related actions.
@@ -16,6 +17,8 @@ use wcf\data\IToggleAction;
  * @method	CaptchaQuestionEditor		getSingleObject()
  */
 class CaptchaQuestionAction extends AbstractDatabaseObjectAction implements IToggleAction {
+	use TDatabaseObjectToggle;
+	
 	/**
 	 * @inheritDoc
 	 */
@@ -25,22 +28,4 @@ class CaptchaQuestionAction extends AbstractDatabaseObjectAction implements ITog
 	 * @inheritDoc
 	 */
 	protected $permissionsUpdate = ['admin.captcha.canManageCaptchaQuestion'];
-	
-	/**
-	 * @inheritDoc
-	 */
-	public function toggle() {
-		foreach ($this->getObjects() as $question) {
-			$question->update([
-				'isDisabled' => $question->isDisabled ? 0 : 1
-			]);
-		}
-	}
-	
-	/**
-	 * @inheritDoc
-	 */
-	public function validateToggle() {
-		parent::validateUpdate();
-	}
 }
