+++ wcfsetup/install/files/lib/data/package/update/server/PackageUpdateServerAction.class.php
@@ -2,6 +2,7 @@
 namespace wcf\data\package\update\server;
 use wcf\data\AbstractDatabaseObjectAction;
 use wcf\data\IToggleAction;
+use wcf\data\TDatabaseObjectToggle;
 
 /**
  * Executes package update server-related actions.
@@ -16,6 +17,8 @@ use wcf\data\IToggleAction;
  * @method	PackageUpdateServerEditor	getSingleObject()
  */
 class PackageUpdateServerAction extends AbstractDatabaseObjectAction implements IToggleAction {
+	use TDatabaseObjectToggle;
+	
 	/**
 	 * @inheritDoc
 	 */
@@ -40,20 +43,4 @@ class PackageUpdateServerAction extends AbstractDatabaseObjectAction implements 
 	 * @inheritDoc
 	 */
 	protected $requireACP = ['create', 'delete', 'toggle', 'update'];
-	
-	/**
-	 * @inheritDoc
-	 */
-	public function validateToggle() {
-		parent::validateUpdate();
-	}
-	
-	/**
-	 * @inheritDoc
-	 */
-	public function toggle() {
-		foreach ($this->getObjects() as $server) {
-			$server->update(['isDisabled' => $server->isDisabled ? 0 : 1]);
-		}
-	}
 }
